requirements.txt
albumentations==1.1.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
conda==23.1.0
conda-content-trust==0.1.3
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
filelock==3.9.0
fonttools==4.38.0
huggingface-hub==0.12.1
idna==3.4
imageio==2.26.0
imgaug==0.4.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
networkx==3
numpy==1.24.2
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
packaging==23
Pillow==9.4.0
pip==22.3.1
pluggy==1.0.0
pycosat==0.6.4
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6
qudida==0.0.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safetensors==0.2.8
scikit-image==0.18.1
scikit-learn==1.2.1
scipy==1.10.1
setuptools==65.6.3
shapely==2.0.1
six==1.16.0
threadpoolctl==3.1.0
tifffile==2023.2.28
toolz==0.12.0
torch==1.8.1+cu101
torchvision==0.9.1+cu101
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
wheel==0.37.1
zipp==3.15.0
zstandard==0.18.0
